feat: Move validation to utilities and add spinner to dropdown
