+++ src/components/LinkChooser/types.ts
@@ -2,7 +2,6 @@
 
 import { CheckboxState } from "@components/Checkbox/Checkbox";
 import { MenuBlock, MenuItemType } from "@components/Dropdown/SelectMenu/SelectMenu";
-import { Validation } from "@components/TextInput";
 import { AriaListBoxOptions } from "@react-aria/listbox";
 import { ComboBoxState } from "@react-stately/combobox";
 import { ListState } from "@react-stately/list";
@@ -19,6 +18,7 @@ import {
     RefObject,
 } from "react";
 import { DoneInvokeEvent, Interpreter } from "xstate";
+import { Validation } from "@utilities/validation";
 
 export type MachineService = Interpreter<
     LinkChooserContext,
@@ -31,13 +31,6 @@ export type MachineService = Interpreter<
         context: LinkChooserContext;
     }
 >;
-
-export const validationClassMap: Record<Validation, string> = {
-    [Validation.Default]: "tw-border-black-20",
-    [Validation.Loading]: "tw-border-black-10",
-    [Validation.Success]: "tw-border-green-50",
-    [Validation.Error]: "tw-border-red-60",
-};
 
 export enum IconLabel {
     Block = "BLOCK",
